feat: add types for comments
