+++ news-api/src/entities/comments/model/index.ts
@@ -1,0 +1,1 @@
+export type { Comments } from './types';
